refactor(scoring): extract shared foundation into admin_scoring_shared.jsx
admin_scoring_modal.jsx was 2643 lines. Move the foundation both editors build
on — pure scoring helpers (fouls/encho/decision/draw logic) plus the small
presentational components (IpponLegend, EnchoControl, DecisionPrompt,
RemainingMatchesPanel, FoulCounter, LineupNameInput, ReasonPrompt, …) — into a
new admin_scoring_shared.jsx. The modal file (now 1949 lines) keeps only the two
stateful editors, imports the foundation, and re-exports the test-facing subset
so `import { … } from './admin_scoring_modal.jsx'` call sites and the vitest
suite are unchanged. No behavior change; 1557 vitest tests green, runtime import
chain verified in-browser.

Refs mp-c2yr.
